Move the character set method to the enigma class adn delete the encryption class
