--- lib/encryption.rb
@@ -1,12 +1,0 @@
-class Encryption
-
-  def initialize(message, key, date)
-  end
-
-  def character_set
-    ("a".."z").to_a << " "
-  end
-
-  
-
-end
