+++ lib/enigma.rb
@@ -3,4 +3,12 @@ class Enigma
   def initialize
   end
 
+  def character_set
+    ("a".."z").to_a << " "
+  end
+
+  def encrypt(message, key, month)
+    hash = {}
+    encryoted
+  end
 end
