+++ spec/enigma_spec.rb
@@ -11,7 +11,14 @@ RSpec.describe Enigma do
   it "can encrypt a message" do
     enigma = Enigma.new
 
-    expect(enigma.encrypt("kyle", "02715", "111122")).to eq("ufjx")
+    expect(enigma.encrypt("kyle", "02715", "111122")).to eq()
+  end
+
+  it 'has a character set' do
+    encryption = Encryption.new("hello", "12345", "111122")
+
+    expect(encryption.character_set.first).to eq("a")
+    expect(encryption.character_set.last).to eq (" ")
   end
 
 end
